[#60]rename: model -> models
