+++ app/visualize/analysis/stmt/expr/parser/constant_expr.py
@@ -1,6 +1,6 @@
 import ast
 
-from app.visualize.analysis.stmt.expr.model.expr_obj import ConstantObj
+from app.visualize.analysis.stmt.expr.models.expr_obj import ConstantObj
 
 
 class ConstantExpr:
