+++ app/visualize/analysis/stmt/expr/parser/list_expr.py
@@ -1,7 +1,7 @@
 from typing import List
 
 from app.visualize.analysis.stmt.expr.expr_util import util
-from app.visualize.analysis.stmt.expr.model.expr_obj import ListObj, ExprObj
+from app.visualize.analysis.stmt.expr.models.expr_obj import ListObj, ExprObj
 
 
 class ListExpr:
