+++ app/visualize/analysis/stmt/expr/parser/name_expr.py
@@ -1,7 +1,7 @@
 import ast
 
 from app.visualize.analysis.element_manager import CodeElementManager
-from app.visualize.analysis.stmt.expr.model.expr_obj import ExprObj, NameObj
+from app.visualize.analysis.stmt.expr.models.expr_obj import ExprObj, NameObj
 
 
 class NameExpr:
