+++ app/visualize/analysis/stmt/parser/assign_stmt.py
@@ -2,8 +2,8 @@ import ast
 
 from app.visualize.analysis.element_manager import CodeElementManager
 from app.visualize.analysis.stmt.expr.expr_traveler import ExprTraveler
-from app.visualize.analysis.stmt.model.assign_stmt_obj import AssignStmtObj
-from app.visualize.analysis.stmt.model.expr_stmt_obj import ExprStmtObj
+from app.visualize.analysis.stmt.models.assign_stmt_obj import AssignStmtObj
+from app.visualize.analysis.stmt.models.expr_stmt_obj import ExprStmtObj
 
 
 class AssignStmt:
