+++ app/visualize/analysis/stmt/parser/if_stmt.py
@@ -2,7 +2,7 @@ import ast
 
 from app.visualize.analysis.element_manager import CodeElementManager
 from app.visualize.analysis.stmt.expr.expr_traveler import ExprTraveler
-from app.visualize.analysis.stmt.model.if_stmt_obj import ElseConditionObj, IfConditionObj, ElifConditionObj
+from app.visualize.analysis.stmt.models.if_stmt_obj import ElseConditionObj, IfConditionObj, ElifConditionObj
 
 
 class IfStmt:
