+++ app/visualize/generator/converter/assign_converter.py
@@ -1,4 +1,4 @@
-from app.visualize.analysis.stmt.model.assign_stmt_obj import AssignStmtObj
+from app.visualize.analysis.stmt.models.assign_stmt_obj import AssignStmtObj
 from app.visualize.generator.highlight.expr_highlight import ExprHighlight
 from app.visualize.generator.highlight.list_highlight import ListHighlight
 from app.visualize.generator.model.assign_viz import AssignViz
