+++ app/visualize/generator/converter_traveler.py
@@ -1,6 +1,6 @@
-from app.visualize.analysis.stmt.model.assign_stmt_obj import AssignStmtObj
-from app.visualize.analysis.stmt.model.for_stmt_obj import ForStmtObj
-from app.visualize.analysis.stmt.model.if_stmt_obj import IfStmtObj
+from app.visualize.analysis.stmt.models.assign_stmt_obj import AssignStmtObj
+from app.visualize.analysis.stmt.models.for_stmt_obj import ForStmtObj
+from app.visualize.analysis.stmt.models.if_stmt_obj import IfStmtObj
 from app.visualize.generator.converter.assign_converter import AssignConverter
 from app.visualize.generator.converter.expr_converter import ExprConverter
 from app.visualize.generator.converter.for_header_converter import ForHeaderConvertor
