+++ tests/visualize/analysis/stmt/expr/parser/test_constant_expr.py
@@ -2,7 +2,7 @@ import ast
 
 import pytest
 
-from app.visualize.analysis.stmt.expr.model.expr_obj import ConstantObj
+from app.visualize.analysis.stmt.expr.models.expr_obj import ConstantObj
 from app.visualize.analysis.stmt.expr.parser.constant_expr import ConstantExpr
