+++ tests/visualize/analysis/stmt/expr/parser/test_name_expr.py
@@ -2,7 +2,7 @@ import ast
 
 import pytest
 
-from app.visualize.analysis.stmt.expr.model.expr_obj import ExprObj, NameObj
+from app.visualize.analysis.stmt.expr.models.expr_obj import ExprObj, NameObj
 from app.visualize.analysis.stmt.expr.parser.name_expr import NameExpr
